Rename sql_langserver to sql_language_server
