+++ py_src/jupyter_lsp/specs/__init__.py
@@ -7,7 +7,7 @@ from .dockerfile_language_server_nodejs import DockerfileLanguageServerNodeJS
 from .javascript_typescript_langserver import JavascriptTypescriptLanguageServer
 from .pyls import PythonLanguageServer
 from .r_languageserver import RLanguageServer
-from .sql_langserver import SQLLanguageServer
+from .sql_language_server import SQLLanguageServer
 from .texlab import Texlab
 from .unified_language_server import UnifiedLanguageServer
 from .vscode_css_languageserver import VSCodeCSSLanguageServer
